requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
better-profanity==0.7.0
colorama==0.4.6
pyTelegramBotAPI==4.24.0